requirements.txt
lxml==6.0.1
python-docx==1.2.0
typing_extensions==4.15.0
fastapi==0.116.1
uvicorn==0.36.0
jinja2==3.1.6
python-multipart==0.0.20
